ai-engine/backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Async tasks
celery==5.3.6
redis==5.0.1

# DAG Orchestration
networkx==3.2.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# APIs externas
google-generativeai==0.3.2

# Utils
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
